chore: send gtag only on first render
move to Misc.jsx and remove GDocAdvertBanner
