+++ static/js/TextColumnBanner.jsx
@@ -1,4 +1,4 @@
-import React, {useState} from "react";
+import React, {useState, useRef, useEffect} from "react";
 import Sefaria  from './sefaria/sefaria';
 import $  from './sefaria/sefariaJquery';
 import {
@@ -114,51 +114,6 @@ const OpenTransBanner = ({ openTranslations }) => {
  * @returns {JSX.Element}
  * @constructor
  */
-
-const GDocAdvertText = () => {
-    const handleLearnMore = () => {
-        gtag('event', 'gdoc_learn_more');
-    }
-    return    <InterfaceText>
-                <EnglishText> Add texts directly to your Google Docs with our <span id="newExtension">new extension</span>! <a href="https://www.sefaria.org/sheets/529099" onClick={handleLearnMore}>Learn more</a></EnglishText>
-                <HebrewText> הוסיפו טקסטים מספריא ישירות לקובץ עם <span id="newExtension">התוסף החדש</span> שלנו! <a href="https://www.sefaria.org/sheets/529099" onClick={handleLearnMore}>למדו עוד</a></HebrewText>
-             </InterfaceText>;
-}
-const GDocInstallLink = 'https://workspace.google.com/marketplace/app/sefaria/849562338091';
-export const GDocAdvertBox = () => {
-    const gdocInstalled = 'gdoc_installed';
-    const handleInstall = () => {
-        cookie(gdocInstalled, JSON.stringify(1), {path: "/"});
-        gtag('event', 'gdoc_installed', {loc: 'Resources Panel'});
-    }
-    return !cookie(gdocInstalled) &&
-            <div className="gDocAdvertBox">
-                <GDocAdvertText/>
-                <div id="installNow"><a href={GDocInstallLink} onClick={handleInstall}>Install Now</a></div>
-            </div>;
-}
-
-export const GDocAdvertBanner = () => {
-    const buttons = [{
-        text: "Install Now",
-        onClick: () => { window.location.href = GDocInstallLink; },
-        sideEffect: "close",
-    }];
-    const onClose = () => {
-        cookie("gdoc_advert_banner_shown", JSON.stringify(1), {path: "/"});
-    };
-
-
-    return (
-      !cookie("gdoc_advert_banner_shown") && document.location.href.includes('/sheets/') ?
-      <div className="gDocAdvertBanner">
-        <TextColumnBanner buttons={buttons} onClose={onClose}>
-            <GDocAdvertText/>
-        </TextColumnBanner>
-      </div> : null
-    );
-};
-
 
 
 /**
